Move models to a data package
Resolves #293
