+++ src/knowlift/__init__.py
@@ -6,25 +6,29 @@
 Bundle core modules whose logic and configuration power this application.
 
 Functions:
-==========
+----------
     create_app: Create and configure a Flask application.
+    get_connection: Get or create a single DB API connection from the pool.
+    close_connection: Return the DB API connection to the pool (Flask teardown).
 
 Modules:
-========
+-------
     db: Store logic that enables database interaction.
-    models: Define entities and relationships among them.
     views: Handle HTTP requests.
 
 Notes:
-======
-    This package bundles the core functionality of this application.
+-----
+    - This package bundles the core functionality of this application.
+    - Db connections are automatically closed via Flask's teardown handlers.
+    - The db engine is stored in Flask's config for application-wide access.
 
 Miscellaneous objects:
-======================
+---------------------
     Except for the public objects exported by this module and their
     public APIs (if applicable), everything else is an implementation
     detail and shouldn't be relied upon, as it may change over time.
 """
+
 
 # Standard Library
 import os
@@ -34,7 +38,7 @@ from logging import config as logging_config
 import flask
 
 # Project specific
-from knowlift import db
+from knowlift import data
 from knowlift import views
 from knowlift.core import settings
 
@@ -65,7 +69,7 @@ def create_app(env=None):
 
     logging_config.dictConfig(app.config['LOGGING_CONFIG'])
 
-    db.init_db(app)
+    data.init_db(app)
 
     app.add_url_rule('/', 'index', views.index)
     app.add_url_rule('/about', 'about', views.about)
@@ -78,5 +82,62 @@ def create_app(env=None):
     app.register_error_handler(404, views.page_not_found)
     app.register_error_handler(500, views.internal_server_error)
 
-    app.teardown_appcontext(db.close_connection)
+    app.teardown_appcontext(close_connection)
     return app
+
+
+def get_connection():
+    """
+    Get or create a single DB API connection from the connection pool.
+
+    This function manages database connections using Flask's application
+    context (flask.g). If a connection doesn't exist in the current context,
+    it creates a new one using the database engine stored in the Flask
+    application configuration.
+
+    Returns:
+        sqlalchemy.Connection:
+            A db connection object that can be used to to execute SQL queries.
+
+    Raises:
+        KeyError:
+            If 'DATABASE_ENGINE' is not found in the Flask app config.
+        RuntimeError:
+            If called outside of a Flask application context.
+
+    Note:
+        The connection is automatically stored in flask.g and will be reused
+        for subsequent calls within the same request context.
+    """
+    if 'db' not in flask.g:
+        engine = flask.current_app.config['DATABASE_ENGINE']
+        flask.g.db = engine.connect()
+        return flask.g.db
+    else:
+        return flask.g.db
+
+
+def close_connection(exception):
+    """
+    Return the underlying DB API connection to the connection pool.
+
+    This function is typically used as a Flask teardown handler to ensure
+    database connections are properly closed at the end of each request,
+    preventing connection leaks.
+
+    Args:
+        exception (Exception | None):
+            An exception that occurred during request processing, if any.
+            This parameter is required by Flask's teardown handler interface
+            but the function handles both success and error cases.
+    """
+    import logging
+    logger = logging.getLogger(__name__)
+    if exception:
+        logger.error(exception)
+
+    db = getattr(flask.g, 'db', None)
+    if db is not None:
+        db.close()
+    else:
+        logger.debug('The database does not exist on the application context.')
